# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JAM CRITTY event registration API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type RegistrationInputGender = typeof RegistrationInputGender[keyof typeof RegistrationInputGender];


export const RegistrationInputGender = {
  female: 'female',
  male: 'male',
} as const;

export interface RegistrationInput {
  /** @minLength 2 */
  full_name: string;
  gender: RegistrationInputGender;
  /** @minLength 7 */
  phone: string;
  instagram_handle: string;
  /** Honeypot field */
  website?: string;
}

export type RegistrationGender = typeof RegistrationGender[keyof typeof RegistrationGender];


export const RegistrationGender = {
  female: 'female',
  male: 'male',
} as const;

export type RegistrationStatus = typeof RegistrationStatus[keyof typeof RegistrationStatus];


export const RegistrationStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
  waitlist: 'waitlist',
} as const;

export type RegistrationPaymentStatus = typeof RegistrationPaymentStatus[keyof typeof RegistrationPaymentStatus];


export const RegistrationPaymentStatus = {
  unpaid: 'unpaid',
  pending: 'pending',
  paid: 'paid',
} as const;

/**
 * @nullable
 */
export type RegistrationPaymentMethod = typeof RegistrationPaymentMethod[keyof typeof RegistrationPaymentMethod] | null;


export const RegistrationPaymentMethod = {
  cash: 'cash',
  ecocash: 'ecocash',
} as const;

export interface Registration {
  id: string;
  full_name: string;
  gender: RegistrationGender;
  phone: string;
  instagram_handle: string;
  status: RegistrationStatus;
  payment_status: RegistrationPaymentStatus;
  /** @nullable */
  payment_method?: RegistrationPaymentMethod;
  notes: string;
  ticket_token: string;
  checked_in: boolean;
  /** @nullable */
  checked_in_at?: string | null;
  created_at: string;
}

export interface RegistrationResult {
  registration: Registration;
  waitlisted: boolean;
}

export type RegistrationUpdateStatus = typeof RegistrationUpdateStatus[keyof typeof RegistrationUpdateStatus];


export const RegistrationUpdateStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
  waitlist: 'waitlist',
} as const;

export type RegistrationUpdatePaymentStatus = typeof RegistrationUpdatePaymentStatus[keyof typeof RegistrationUpdatePaymentStatus];


export const RegistrationUpdatePaymentStatus = {
  unpaid: 'unpaid',
  pending: 'pending',
  paid: 'paid',
} as const;

export type RegistrationUpdatePaymentMethod = typeof RegistrationUpdatePaymentMethod[keyof typeof RegistrationUpdatePaymentMethod];


export const RegistrationUpdatePaymentMethod = {
  cash: 'cash',
  ecocash: 'ecocash',
} as const;

export interface RegistrationUpdate {
  status?: RegistrationUpdateStatus;
  payment_status?: RegistrationUpdatePaymentStatus;
  payment_method?: RegistrationUpdatePaymentMethod;
  notes?: string;
  checked_in?: boolean;
}

export interface Stats {
  femaleCount: number;
  maleCount: number;
  totalCount: number;
  maxFemale: number;
  maxMale: number;
  maxTotal: number;
  femaleSpotsLeft: number;
  maleSpotsLeft: number;
  totalSpotsLeft: number;
  femaleFull: boolean;
  maleFull: boolean;
  isFull: boolean;
  femaleWaitlistCount: number;
  maleWaitlistCount: number;
  checkedInCount: number;
  paidCount: number;
}

export interface TicketLookupInput {
  token: string;
}

export interface CheckinInput {
  id: string;
  collect_cash?: boolean;
}

export type WalkupInputGender = typeof WalkupInputGender[keyof typeof WalkupInputGender];


export const WalkupInputGender = {
  female: 'female',
  male: 'male',
} as const;

export interface WalkupInput {
  /** @minLength 2 */
  full_name: string;
  gender: WalkupInputGender;
}

export interface EventSettings {
  id: number;
  max_total: number;
  max_female: number;
  max_male: number;
  updated_at: string;
}

export interface EventSettingsInput {
  /** @minimum 0 */
  max_total: number;
  /** @minimum 0 */
  max_female: number;
  /** @minimum 0 */
  max_male: number;
}

export interface ErrorResponse {
  error: string;
}