# lib/api-zod/src/generated/types/checkinInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JAM CRITTY event registration API
 * OpenAPI spec version: 0.1.0
 */

export interface CheckinInput {
  id: string;
  collect_cash?: boolean;
}