# lib/api-zod/src/generated/types/eventSettings.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JAM CRITTY event registration API
 * OpenAPI spec version: 0.1.0
 */

export interface EventSettings {
  id: number;
  max_total: number;
  max_female: number;
  max_male: number;
  updated_at: string;
}