# lib/api-zod/src/generated/types/eventSettingsInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JAM CRITTY event registration API
 * OpenAPI spec version: 0.1.0
 */

export interface EventSettingsInput {
  /** @minimum 0 */
  max_total: number;
  /** @minimum 0 */
  max_female: number;
  /** @minimum 0 */
  max_male: number;
}