# lib/api-zod/src/generated/types/registration.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JAM CRITTY event registration API
 * OpenAPI spec version: 0.1.0
 */
import type { RegistrationGender } from './registrationGender';
import type { RegistrationPaymentMethod } from './registrationPaymentMethod';
import type { RegistrationPaymentStatus } from './registrationPaymentStatus';
import type { RegistrationStatus } from './registrationStatus';

export interface Registration {
  id: string;
  full_name: string;
  gender: RegistrationGender;
  phone: string;
  instagram_handle: string;
  status: RegistrationStatus;
  payment_status: RegistrationPaymentStatus;
  /** @nullable */
  payment_method?: RegistrationPaymentMethod;
  notes: string;
  ticket_token: string;
  checked_in: boolean;
  /** @nullable */
  checked_in_at?: string | null;
  created_at: string;
}