# lib/api-zod/src/generated/types/registrationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JAM CRITTY event registration API
 * OpenAPI spec version: 0.1.0
 */
import type { RegistrationInputGender } from './registrationInputGender';

export interface RegistrationInput {
  /** @minLength 2 */
  full_name: string;
  gender: RegistrationInputGender;
  /** @minLength 7 */
  phone: string;
  instagram_handle: string;
  /** Honeypot field */
  website?: string;
}