# lib/api-zod/src/generated/types/registrationResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JAM CRITTY event registration API
 * OpenAPI spec version: 0.1.0
 */
import type { Registration } from './registration';

export interface RegistrationResult {
  registration: Registration;
  waitlisted: boolean;
}