# lib/api-zod/src/generated/types/registrationUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JAM CRITTY event registration API
 * OpenAPI spec version: 0.1.0
 */
import type { RegistrationUpdatePaymentMethod } from './registrationUpdatePaymentMethod';
import type { RegistrationUpdatePaymentStatus } from './registrationUpdatePaymentStatus';
import type { RegistrationUpdateStatus } from './registrationUpdateStatus';

export interface RegistrationUpdate {
  status?: RegistrationUpdateStatus;
  payment_status?: RegistrationUpdatePaymentStatus;
  payment_method?: RegistrationUpdatePaymentMethod;
  notes?: string;
  checked_in?: boolean;
}