# lib/api-zod/src/generated/types/stats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JAM CRITTY event registration API
 * OpenAPI spec version: 0.1.0
 */

export interface Stats {
  femaleCount: number;
  maleCount: number;
  totalCount: number;
  maxFemale: number;
  maxMale: number;
  maxTotal: number;
  femaleSpotsLeft: number;
  maleSpotsLeft: number;
  totalSpotsLeft: number;
  femaleFull: boolean;
  maleFull: boolean;
  isFull: boolean;
  femaleWaitlistCount: number;
  maleWaitlistCount: number;
  checkedInCount: number;
  paidCount: number;
}